refactor: secretKey 숨기기
