+++ src/main/java/gift/member/oauth/KakaoTokenResponse.java
@@ -3,6 +3,8 @@ package gift.member.oauth;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record KakaoTokenResponse(@JsonProperty("access_token") String accessToken,
-                                 @JsonProperty("refresh_token") String refreshToken) {
+                                 @JsonProperty("expires_in") Integer accessTokenExpiresIn,
+                                 @JsonProperty("refresh_token") String refreshToken,
+                                 @JsonProperty("refresh_token_expires_in") Integer refreshTokenExpiresIn) {
 
 }
